backend/pyproject.toml
[tool.poetry]
name = "autosinan"
version = "0.1.0"
description = ""
authors = ["sr-pato <felipe.adeildo@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
dbfread = "^2.0.7"
icecream = "^2.1.3"
toml = "^0.10.2"
xlsxwriter = "^3.2.0"
openpyxl = "^3.1.4"
fastapi = "^0.112.1"
pydantic-settings = "^2.4.0"
uvicorn = "^0.30.6"
python-jose = "^3.3.0"
prisma = "^0.15.0"
python-multipart = "^0.0.9"
bcrypt = "^4.2.0"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
ipython = "^8.22.1"
ruff = "^0.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
